backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
SQLAlchemy==2.0.34
pymysql==1.1.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart